+++ src/main/java/com/map/gaja/client/infrastructure/geocode/exception/NotExcelUploadException.java
@@ -1,4 +1,4 @@
-package com.map.gaja.client.domain.service.geocode.exception;
+package com.map.gaja.client.infrastructure.geocode.exception;
 
 import com.map.gaja.global.exception.BusinessException;
 import org.springframework.http.HttpStatus;
